requirements.txt
datasets==4.6.1
numpy==1.26.4
opencv-python==4.11.0.86
Pillow==10.4.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1

